move dummy app to tests to avoid installing dependencies in production
