+++ botbot_plugins/base.py
@@ -1,9 +1,3 @@
-from cmd import Cmd
-import re
-import sys
-
-import fakeredis
-
 class PrivateMessage(object):
     """
     A holder object for sending a private message.
@@ -81,210 +75,3 @@ class BasePlugin(object):
         """
         ukey = self._unique_key(key)
         return self.app.storage.incr(ukey)
-
-
-class DummyLine(object):
-    """
-    All the methods and data necessary for a plugin to act on a line
-    """
-    def __init__(self, packet):
-        self.text = packet['text']
-        self.full_text = packet['text']
-        self.user = packet.get('User', 'repl_user')
-        self._channel_name = packet.get('Channel', '#dummy-channel')
-        self.is_direct_message = self.check_direct_message()
-        self._command = packet.get('Command', 'PRIVMSG')
-        self._is_message = self._command == 'PRIVMSG'
-
-    def check_direct_message(self):
-        """Are you addressing the bot?"""
-        if self.text.startswith('@'):
-            self.text = self.text[1:]
-            return True
-        return False
-
-    def __str__(self):
-        return self.full_text
-
-    def __repr__(self):
-        return str(self)
-
-
-class DummyPrivateMessage(object):
-    def __init__(self, nick, msg):
-        self.nick = nick
-        self.msg = msg
-
-
-REPL_INTRO = """
-#########################
-The BotBot.me plugin repl
-#########################
-
-Type a line to see how the plugins respond.
-Prefix the line with `@` to send a direct message to the bot. Example:
-    @ping
-
-To configure a plugin, use `!!{plugin_name}:{field_name}={value}`. Example:
-    !!github:organization=lincolnloop
-"""
-
-
-class DummyApp(Cmd):
-    """
-    Registration and routing for plugins
-    """
-    prompt = '(repl_user) '
-    intro = REPL_INTRO
-    use_raw_input = False
-
-    def __init__(self, *args, **kwargs):
-        # Cmd is an old-style class, super doesn't work
-        # super(DummyApp, self).__init__(*args, **kwargs)
-        self.responses = []
-        self.storage = fakeredis.FakeStrictRedis()
-        self.routers = {
-            "firehose": Router("firehose"),
-            "messages": Router("messages"),
-            "mentions": Router("mentions"),
-            "commands": Router("commands"),
-            "regex_commands": Router("regex_commands")
-        }
-        self.plugin_configs = {}
-        if 'test_plugin' in kwargs:
-            self.test_mode = True
-            self.test_plugin = kwargs['test_plugin']
-            self.register(kwargs['test_plugin'])
-            del(kwargs['test_plugin'])
-        else:
-            self.test_mode = False
-
-        if "command_prefix" in kwargs:
-            self.command_prefix = kwargs["command_prefix"]
-            del(kwargs["command_prefix"])
-        else:
-            self.command_prefix = "!"
-
-        Cmd.__init__(self, *args, **kwargs)
-
-    def register(self, plugin):
-        """
-        Introspects the Plugin class instance provided for methods
-        that need to be registered with the internal app routers.
-        """
-        plugin.app = self
-        for key in dir(plugin):
-            attr = getattr(plugin, key)
-            if not key.startswith('__') and getattr(attr, 'route_rule', None):
-                router_name = attr.route_rule[0]
-                if router_name in [r.name for _, r in self.routers.items()]:
-                    rule = attr.route_rule[1]
-
-                    self.output('Route {}: {} ({}, {})'.format(router_name,
-                                                                plugin.slug, key,
-                                                                rule))
-                    self.routers[router_name].plugins.setdefault(
-                        plugin.slug, []).append((rule, attr))
-                    # Setup the plugin config
-                    if (plugin.config_class and
-                            plugin.slug not in self.plugin_configs):
-                        self.plugin_configs[plugin.slug] = plugin.config_class()
-
-    def output(self, text):
-        """Print text to stdout for repl. No-op for tests"""
-        if not self.test_mode:
-            print(text)
-
-    def set_config(self, plugin_slug, fields_dict):
-        """Manually set a plugin config. Used for testing"""
-        self.plugin_configs[plugin_slug].fields.update(fields_dict)
-
-    def respond(self, text, **kwargs):
-        """Listens for incoming messages"""
-        if text.startswith('!!'):
-            return self.do_config(text)
-        self.responses = []
-
-        packet = {'text': text}
-        packet.update(kwargs)
-
-        line = DummyLine(packet)
-        self.dispatch(line)
-        if self.test_mode:
-            return self.responses
-
-    # Cmd sends all text to the `default` method
-    default = respond
-
-    def do_EOF(self, arg):
-        """Kill cmdloop on CTRL-d"""
-        print("\nGoodbye")
-        sys.exit()
-
-    def do_config(self, arg):
-        """Handles lines prefaced with `!!`. Used to set config values"""
-        try:
-            plugin_slug, eq = arg[2:].split(':', 1)
-            field, value = eq.split('=', 1)
-        except ValueError:
-            print("Bad config format. {plugin_slug}:{field_name}={value}")
-            return
-        if plugin_slug not in self.plugin_configs:
-            print(('No config defined for "{0}"'.format(plugin_slug)))
-            return
-        if field not in self.plugin_configs[plugin_slug].fields:
-            print(('Field "{0}" is not defined for "{1}"'.format(field,
-                                                                plugin_slug)))
-            return
-        self.plugin_configs[plugin_slug].fields[field] = value
-        print('Config saved.')
-
-    def dispatch(self, line):
-        """Given a line, dispatch it to the right function(s)"""
-        self.check_routes_for_matches(line, self.routers["messages"])
-
-        if line.is_direct_message:
-            self.check_routes_for_matches(line, self.routers["mentions"])
-
-        if line.text.startswith(self.command_prefix):
-            self.check_routes_for_matches(line, self.routers["commands"])
-            self.check_routes_for_matches(line, self.routers["regex_commands"])
-
-        self.check_routes_for_matches(line, self.routers["firehose"])
-
-    def check_routes_for_matches(self, line, router):
-        """Checks if line matches the routes' rules and calls functions"""
-        for _, route_list in router.plugins.items():
-            for rule, func in route_list:
-                if router.name == "commands":
-                    cmd = rule
-                    args = line.text.split()
-                    if args[0] == self.command_prefix + cmd:
-                        arg_dict = {"args": args[1:]}
-                        self.run_plugin(line, func, arg_dict)
-
-                elif router.name == "regex_commands":
-                    cmd = rule[0]
-                    regex = rule[1]
-                    args = line.text.split()
-                    if args[0] == self.command_prefix + cmd:
-                        linetext = " ".join(args[1:])
-                        match = re.match(regex, linetext, re.IGNORECASE)
-                        if match:
-                            self.run_plugin(line, func, match.groupdict())
-                else:
-                    match = re.match(rule, line.text, re.IGNORECASE)
-                    if match:
-                        self.run_plugin(line, func, match.groupdict())
-
-    def run_plugin(self, line, func, arg_dict):
-        response = func(line, **arg_dict)
-        if response:
-            if isinstance(response, PrivateMessage):
-                self.responses.append(response.msg)
-                self.output('[o__o]: ' + response.msg)
-            else:
-                self.responses.append(response)
-                self.output('[o__o]: ' + response)
-
-app = DummyApp()
