+++ botbot_plugins/tests/test_jira.py
@@ -6,7 +6,7 @@ import json
 import time
 from unittest.mock import patch, call
 import requests
-from botbot_plugins.base import DummyApp
+from botbot_plugins.tests.dummy import DummyApp
 from botbot_plugins.plugins import jira
