+++ botbot_plugins/tests/test_metabrain.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 import pytest
-from botbot_plugins.base import DummyApp
+from botbot_plugins.tests.dummy import DummyApp
 from botbot_plugins.plugins import metabrain
